Credit-Card-Reader-Based-on-Face-Recognition
In this study, a new system is proposed for the payment stage of online shopping. The study consists of two main parts. The first step is to scan the credit card with the webcam and extract the card number. The second stage is to allow payment by facial recognition. This study aims to provide speed and convenience to the user and to provide security while providing these, and even to increase security. Face recognition consists of 3 stages in itself. The first stage of face recognition is the stage where 30 photos are taken and stored as a dataset to introduce people to the system. The second stage is the training of the photos stored as the dataset by the selected algorithm. The third stage, the final stage, is the webcam recognition of the face of the user authenticating via the webcam. The LBHP algorithm with 92% performance was chosen for face recognition part.